+++ aio_request/__init__.py
@@ -23,8 +23,9 @@ from .endpoint_provider import DelegateEndpointProvider, EndpointProvider, Stati
 from .pipeline import BypassModule, LowTimeoutModule, NextModuleFunc, RequestModule, TransportModule, build_pipeline
 from .priority import Priority
 from .request import (
+    AsyncRequestEnricher,
+    DeprecatedAsyncRequestEnricher,
     RequestEnricher,
-    SimpleRequestEnricher,
     delete,
     get,
     patch,
@@ -110,8 +111,9 @@ __all__: tuple[str, ...] = (
     "post_json",
     "put",
     "put_json",
-    "SimpleRequestEnricher",
     "RequestEnricher",
+    "AsyncRequestEnricher",
+    "DeprecatedAsyncRequestEnricher",
     # request_strategy.py
     "MethodBasedStrategy",
     "ParallelRequestStrategy",