+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 ## v0.1.34 (2024-11-05)
 
 * [Try to get metrics provider from transport in setup_v2 if no metrics provider is passed](https://github.com/anna-money/aio-request/pull/280)
-π
+
 
 ## v0.1.33 (2024-10-29)
 
@@ -19,6 +19,7 @@
 
 
 ## v0.1.32 (2024-10-18)
+
 * [Endpoint provider](https://github.com/anna-money/aio-request/pull/270)